+++ lib/kitchen/terraform/transport.rb
@@ -14,24 +14,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require "kitchen"
-require "kitchen/terraform/verify_version_rescue_strategy/permissive"
-require "support/kitchen/logger_context"
-
-::RSpec.describe ::Kitchen::Terraform::VerifyVersionRescueStrategy::Permissive do
-  subject do
-    described_class.new logger: logger
-  end
-
-  include_context "Kitchen::Logger"
-
-  describe "#call" do
-    specify "should warn the user that the Terraform client version is unsupported" do
-      expect(logger).to receive :warn
-    end
-
-    after do
-      subject.call
+module Kitchen
+  module Terraform
+    # Transport is the namespace for transport strategies.
+    module Transport
     end
   end
 end