+++ spec/support/kitchen/logger_context.rb
@@ -14,24 +14,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require "kitchen"
-require "kitchen/terraform/verify_version_rescue_strategy/permissive"
-require "support/kitchen/logger_context"
+require "kitchen/logger"
 
-::RSpec.describe ::Kitchen::Terraform::VerifyVersionRescueStrategy::Permissive do
-  subject do
-    described_class.new logger: logger
-  end
-
-  include_context "Kitchen::Logger"
-
-  describe "#call" do
-    specify "should warn the user that the Terraform client version is unsupported" do
-      expect(logger).to receive :warn
-    end
-
-    after do
-      subject.call
-    end
+::RSpec.shared_context "Kitchen::Logger" do
+  let :logger do
+    ::Kitchen::Logger.new stdout: ::File.open(::File::NULL, "w")
   end
 end